+++ src/Sitecore.Rocks.Server/ServerHost.cs
@@ -1,23 +1,14 @@
 ﻿// © 2015-2016 Sitecore Corporation A/S. All rights reserved.
 
 using Sitecore.Rocks.Server.Extensibility;
-using Sitecore.Rocks.Server.Hosts;
 
 namespace Sitecore
 {
     public static class ServerHost
     {
-        private static AppVersionHost appVersionHost;
-
         private static EnvHost env;
 
         private static ExtensibilityHost extensibility;
-
-        [NotNull]
-        public static AppVersionHost AppVersion
-        {
-            get { return appVersionHost ?? (appVersionHost = new AppVersionHost()); }
-        }
 
         [NotNull]
         public static EnvHost Env